feat(js): datatables view upgrade;
- Upgraded vendor distros for datatables view.
- Started recoding the custom module to functional JS code.
