+++ ckanext/datatablesview/assets/webassets.yml
@@ -1,20 +1,17 @@
 main-css:
   output: ckanext-datatablesview/%(version)s_datatablesview.css
   contents:
-    - vendor/DataTables/datatables.css
-    - vendor/DataTables/datatables.mark.css
-    - datatables_view.css
+    - vendor/DataTables/datatables.min.css
+    - datatablesview.css
 main-js:
   output: ckanext-datatablesview/%(version)s_datatablesview.js
   extra:
     preload:
       - base/main
   contents:
-    - vendor/DataTables/datatables.js
-    - vendor/DataTables/datatables.mark.js
-    - vendor/DataTables/jquery.mark.js
-    - vendor/DataTables/dataTables.scrollResize.js
-    - vendor/DataTables/datetime.js
+    - vendor/DataTables/datatables.min.js
+    - vendor/DataTables/datatables.datetime-moment.js
+    - vendor/DataTables/datatables.mark.min.js
     - datatablesview.js
 form-js:
   output: ckanext-datatablesview/%(version)s_datatablesview_form.js
