--- ckanext/datatablesview/public/datatablesview/i18n
@@ -1,1 +1,0 @@
-../../assets/vendor/DataTables/i18n
